feat: add data transforms and basis_scalar to modelling.
API BREAKING: breaks modelling API a little.
